Check whether clusters are Neun+ or Neun-

This function is very specific to brain cell/nuclear data where Neun is used as a stain for neuronal vs. non-neuronal identity. This function determine's whether a cluster is predicted to be neuronal or non-neuronal based on expression of Neun.

check_neun(
  anno,
  cluster,
  neun.thresh = 0.5,
  neun.colname = "facs_population_plan",
  neun.val = "NeuN-pos"
)

Arguments

anno

anno dataframe which must include column name listed in `neun.colname`

cluster

cluster labels for all cells along with sample_id as their names

neun.thresh

fraction of cells expressing NeuN to be considered NeuN positive (default is 0.5)

neun.colname

column name in `anno` with the Neun information

neun.val

value corresponding to non-neuronal marker in neun.colname in anno

Value

returns all the clusters and true/false of whether they are Neun positive