Convert paired cluster comparison values to a matrix

Convert paired cluster comparison values to a matrix

convert_pair_matrix(pair.num, l = NULL, directed = FALSE)

Arguments

pair.num
a named numeric vector of values. Names correspond to compared elements separted by "_", e.g. "c1_c2", "c23_c59"

l
labels for columns. Default is NULL, which will compute them from names(pair.num)

directed
If FALSE (default), the first value in each pair will specify used as columns, with the second as rows. If TRUE, first values will be rows, and second will be columns.

Value

a matrix containing values from pair.num, and named for each element separated by "_" in names(pair.num)

Examples

pair_values <- seq(1,27,3)
names(pair_values) <- paste(rep(letters[1:3], each = 3), rep(letters[1:3], 3), sep = "_")
pair_values

#> a_a a_b a_c b_a b_b b_c c_a c_b c_c 
#>   1   4   7  10  13  16  19  22  25 

pair_matrix <- convert_pair_matrix(pair_values, directed = FALSE)
pair_matrix

#>   a  b  c
#> a 1 10 19
#> b 4 13 22
#> c 7 16 25

pair_matrix <- convert_pair_matrix(pair_values, directed = TRUE)
pair_matrix

#>    a  b  c
#> a  1  4  7
#> b 10 13 16
#> c 19 22 25