Perform pairwise differential gene expression tests between main pairs of clusters in parallel

Perform pairwise differential gene expression tests between main pairs of clusters in parallel

de_selected_pairs(norm.dat, cl, pairs, method = "limma", low.th = 1,
  cl.present = NULL, use.voom = FALSE, counts = NULL, mc.cores = 1)

Arguments

norm.dat
a normalized data matrix.

cl
a cluster factor object.

pairs
A 2-column matrix of cluster pairs.

method
Either "limma" or "chisq".

low.th
The minimum expression value used to filter for expressed genes.

cl.present
A matrix of proportions of cells in each cluster with gene detection. Can be generated with get_cl_props(). Default is NULL (will be generated).

use.voom
Logical, whether or not to use voom() for limma calculations. Default is FALSE.

counts
A matrix of raw count data for each cell. Required if use.voom is TRUE. Default is NULL.

mc.cores
A number indicating how many processor cores to use for parallelization.