Identify high variability genes using Brennecke's method

This function is a wrapper around compute_vg_stats(), which computes variability statistics, and plot_vg(), which generates multiple plots for those results.

find_vg(
  dat,
  rescaled = FALSE,
  plot_file = NULL,
  return_type = "data",
  verbose = FALSE
)

Arguments

dat

A matrix or dgCMatrix with samples as columns and genes as rows

rescaled

logical - whether the data matrix should be normalized by columns). Assume input data have already been normalized

plot_file

(optional) A file to use to output diagnostic plots. Default is NULL.

return_type

What kind of objects to return. Default is "data", which will return only a data.frame with results for each gene. Other options are "plots", which will return a list containing plot objects, and "both" which will return a list with the statistics (gene_var_stats) and plots.

verbose

logical - whether or not to display status messages

Value

Either a data.frame or a list (see return_type parameter).