Find specific marker gene for each cluster

This function identifies the top marker gene (or "none") for each cluster based on the cluster medians and proportions of cells expressing a gene and based on various parameters.

getTopMarkersByPropNew(propExpr, medianExpr, propDiff = 0,
  propMin = 0.5, medianFC = 1, excludeGenes = NULL,
  sortByMedian = TRUE)

Arguments

propExpr
matrix of proportions of cells expressing a gene in each cluster (genes=rows, clusters=columns)

medianExpr
matrix of median expression per cluster (genes=rows, clusters=columns)

propDiff
Must have difference in proportion higher than this value in "on" cluster compared with each other cluster

propMin
Must have higher proportion in "on" cluster

medianFC
Must have median fold change greater than this value in "on" group vs. each other cluster

excludeGenes
Genes exlcuded from marker consideration (NULL by default)

sortByMedian
Should genes passing all filters be prioritized by median fold change (TRUE, default) or by difference in proportion between clusters (FALSE)

Value

a vector of the top marker gene per cluster (or "none" for clusters with none)

Details

Note: this could potentially be replaced by an existing hicat function for identifying marker genes. This function is fast and chooses 'reasonable" genes.