Iterative consensus clustering

Iterative consensus clustering

iter_consensus_clust(
  cl.list,
  co.ratio = NULL,
  cl.mat = NULL,
  norm.dat,
  select.cells = names(cl.list[[1]]),
  diff.th = 0.25,
  prefix = NULL,
  method = c("auto", "louvain", "ward.D"),
  verbose = FALSE,
  de.param = de.param,
  max.cl.size = 300,
  result = NULL,
  split.size = de.param$min.cells * 2,
  merge.type = c("undirectional", "directional")
)

Arguments

cl.list

The list of subsampled clustering results.

co.ratio

cell cell co-clustering matrix

norm.dat

The log2 transformed normalzied expression matrix

select.cells

Cells to be clustered

diff.th

The difference of co-clustering probablities for splitting a cluster.

prefix

Default NULL.

method

Clustering methods. Default "auto"

verbose

Default FALSE

de.param

Differentiall expressed genes criteria for merging clusters

result

Pre-computed clustering results used for further splitting. Default NULL.

merge.type

Determine if the DE gene score threshold should be applied to combined de.score, or de.score for up and down directions separately.

Value

A list with cluster membership, and top pairwise marker genes.