Refine clusters

Refine clusters

refine_cl(
  cl,
  co.ratio = NULL,
  cl.mat = NULL,
  confusion.th = 0.6,
  min.cells = 4,
  niter = 50,
  tol.th = 0.02,
  verbose = 0
)

Arguments

cl

Cluster membership cluster

co.ratio

cell-cell co-clustering matrix.

cl.mat

cell-cluster matrix collected from bootstrapping iterations. Either co.ratio or cl.mat should not be NULL.

confusion.th

Clusters with average confusion score greater than this threshold will be removed. Cells in this cluster will be re-distributed to other most likely clusters.

min.cells

Clusters with fewer than this many cells will be removed. Cells in this cluster will be re-distributed to other most likely clusters.

niter

maxmimal mumber of refinement iterations.

tol.th

If improvement is smaller than this threshold, terminate refinement step.

verbose

If true, print out step-by-step improvement.

Author

Zizhen Yao